Conflicts:
	plugin/IEHostWindow.cpp

+++ plugin/IEHostWindow.cpp
@@ -1500,13 +1500,13 @@ void CIEHostWindow::FBFindAgainInternal(bool backwards, bool norecur, bool nosel
 					m_lFBCurrentDoc--; 
 					if (m_lFBCurrentDoc < 0)
 					{
-						m_lFBCurrentDoc = (long)(m_vFBDocs.size() - 1);
+						m_lFBCurrentDoc = static_cast<long>(m_vFBDocs.size()) - 1;
 						m_bFBCrossHead = true;
 					}
 				} else
 				{
 					m_lFBCurrentDoc++;
-					if (m_lFBCurrentDoc == m_vFBDocs.size())
+					if (m_lFBCurrentDoc ==  static_cast<long>(m_vFBDocs.size()))
 					{
 						m_lFBCurrentDoc = 0;
 						m_bFBCrossTail = true;
@@ -1632,7 +1632,7 @@ void CIEHostWindow::FBHighlightAll()
 	if (m_strFBText.GetLength() == 0) return;
 
 	long lOriginalIndex = m_lFBCurrentDoc;
-	for (m_lFBCurrentDoc = 0; m_lFBCurrentDoc < (long)m_vFBDocs.size(); m_lFBCurrentDoc++)
+	for (m_lFBCurrentDoc = 0; m_lFBCurrentDoc <  static_cast<long>(m_vFBDocs.size()); m_lFBCurrentDoc++)
 	{
 		FBDocFindStatus& dfs = FBGetCurrentDocStatus();
 		long tmp;
